Add GitHub Actions workflow for building and releasing JAR
This workflow automates the building and releasing of a JAR file for the Master-Hunter project when a tag is pushed or manually triggered. It includes steps for checking out the code, setting up Java and Gradle, building the JAR, extracting version information, creating a release, and uploading the JAR as an artifact.